wpandas refactor
