+++ anesthetic/_version.py.orig
@@ -1,0 +1,5 @@
+<<<<<<< HEAD
+__version__ = '2.0.0b26'
+=======
+__version__ = '2.0.0b25'
+>>>>>>> master
