+++ tests/test_labelled_pandas.py
@@ -1,6 +1,6 @@
 import numpy as np
 from numpy.testing import assert_array_equal
-from anesthetic.labelled_pandas import LabelledSeries, LabelledDataFrame
+from lpandas import LabelledSeries, LabelledDataFrame
 from pandas import Series, DataFrame, MultiIndex
 import pandas.testing
 import pytest
